+++ ios/MullvadRESTTests/MullvadApiTests.swift
@@ -35,7 +35,7 @@ class MullvadApiTests: XCTestCase {
                 address: .ipv4(.loopback),
                 port: 1080,
                 password: "123",
-                cipher: CipherIdentifiers.CHACHA20.description
+                cipher: "aes-128-cfb"
             ))
 
         let accessMethodsRepository = AccessMethodRepositoryStub.stub