+++ ios/MullvadVPN/Coordinators/Settings/APIAccess/List/ListAccessViewModelBridge.swift
@@ -6,14 +6,27 @@ class ListAccessViewModelBridge: ListAccessViewModel {
 
     @Published var items: [ListAccessMethodItem] = []
     @Published var itemInUse: ListAccessMethodItem?
+
     init(
         interactor: ListAccessMethodInteractorProtocol,
         delegate: ListAccessMethodViewControllerDelegate?
     ) {
         self.interactor = interactor
         self.delegate = delegate
+
         interactor.itemsPublisher.assign(to: &$items)
         interactor.itemInUsePublisher.assign(to: &$itemInUse)
+    }
+
+    func cipherIsValid(for item: ListAccessMethodItem) -> Bool {
+        let ciphers = interactor.shadowsocksCiphers
+        let method = interactor.accessMethod(by: item.id)
+
+        return if case .shadowsocks(let config) = method?.proxyConfiguration {
+            ciphers.contains(config.cipher)
+        } else {
+            true
+        }
     }
 
     func addNewMethod() {