+++ ios/CHANGELOG.md
@@ -31,6 +31,7 @@ Line wrap the file at 100 chars.                                              Th
 - Add recent connections in the Select location view.
 - Show disabled servers in location view.
 - Add ability which types of local notifications are delivered.
+- Remove invalid Shadowsocks ciphers.
 
 [TunnelCrack]: https://tunnelcrack.mathyvanhoef.com/